requirements.txt
onvif-zeep
PyYAML>=6.0
